Added function to Indexer API to return image key types before attempting to grab images that may result in a HTTP 404 error code.
Adding in MySQL functionality for larger databases that may perform
poorly using SQLite.

Refactored how show statistics are gathered.

+++ sickrage/core/databases/main/db_repository/versions/002_Add_Last_Backlog_Search_Column_To_TVShow_Table.py
@@ -4,11 +4,13 @@ from sqlalchemy import *
 def upgrade(migrate_engine):
     meta = MetaData(bind=migrate_engine)
     tv_shows = Table('tv_shows', meta, autoload=True)
-    last_backlog_search = Column('last_backlog_search', Integer, default=0)
-    last_backlog_search.create(tv_shows)
+    if not hasattr(tv_shows.c, 'last_backlog_search'):
+        last_backlog_search = Column('last_backlog_search', Integer, default=0)
+        last_backlog_search.create(tv_shows)
 
 
 def downgrade(migrate_engine):
     meta = MetaData(bind=migrate_engine)
     tv_shows = Table('tv_shows', meta, autoload=True)
-    tv_shows.c.last_backlog_search.drop()
+    if hasattr(tv_shows.c, 'last_backlog_search'):
+        tv_shows.c.last_backlog_search.drop()
